fix for attempting to decrypt a DB which is not encrypted
this situation can occur if a non encrypted DB was then set as encrypted using the mode flag, but the data remains the same. Whether in encrypt mode or not, if we already have a non encrypted json string, then just return that.

When we later make a change to the DB data, the mode will then cause encryption if necessary.